+++ src/service/search/cluster/mod.rs
@@ -981,7 +981,7 @@ async fn get_file_list_by_inverted_index(
 
     let fts_condition = terms
         .iter()
-        .map(|x| format!("term LIKE '{x}%'"))
+        .map(|x| format!("term LIKE '%{x}%'"))
         .collect::<Vec<_>>()
         .join(" OR ");
     let fts_condition = if fts_condition.is_empty() {
